This works with the sequence number
Hey, here's what I've been working on today. This one will work with the sequence number. It works most of the time but sometimes it enters an endless loop that I have not been able to fix yet. Also, I'm sort of skipping the random number picker step for now. Hope this helps a bit. I'll be around tomorrow to work on it as well.